requirement.txt
pypdf
langchain
openai 
langchain-community
langchain_openai
python-dotenv
chromadb